Reorganize actor utils
Create code.ts

Create generic.ts

Update generic.ts

Create state.ts

Delete utils.ts

Linter fixes

Update index.spec.ts
