+++ packages/filecoin-actor-utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
-export * from './utils'
+export * from './utils/code'
+export * from './utils/state'
